use float suffixes

+++ synfig-studio/src/gui/widgets/widget_colorslider.cpp
@@ -155,7 +155,7 @@ ColorSlider::get_amount(Type type, const synfig::Color& color)
 	case TYPE_V: return color.get_v()+0.5f;
 	case TYPE_HUE: {
 		float amount = Angle::rot(color.get_uv_angle()).get();
-		return amount - floor(amount);
+		return amount - floorf(amount);
 	}
 	case TYPE_SAT: return color.get_s()*2.0f;
 	case TYPE_A: return color.get_a();
@@ -361,10 +361,10 @@ ColorSlider::on_event(GdkEvent *event)
 			case TYPE_G: amount=color.get_g(); break;
 			case TYPE_B: amount=color.get_b(); break;
 			case TYPE_Y: amount=color.get_y(); break;
-			case TYPE_U: amount=color.get_u()+0.5; break;
-			case TYPE_V: amount=color.get_v()+0.5; break;
-			case TYPE_HUE: amount=Angle::rot(color.get_uv_angle()).get(); amount-=floor(amount); break;
-			case TYPE_SAT: amount=color.get_s()*2.0; break;
+			case TYPE_U: amount=color.get_u()+0.5f; break;
+			case TYPE_V: amount=color.get_v()+0.5f; break;
+			case TYPE_HUE: amount=Angle::rot(color.get_uv_angle()).get(); amount-=floorf(amount); break;
+			case TYPE_SAT: amount=color.get_s()*2.0f; break;
 			case TYPE_A: amount=color.get_a(); break;
 			default: amount=0; break;
 		}
@@ -372,11 +372,11 @@ ColorSlider::on_event(GdkEvent *event)
 		switch(event->scroll.direction){
 			case GDK_SCROLL_UP:
 			case GDK_SCROLL_RIGHT:
-				x+=1.0;
+				x+=1.0f;
 				break;
 			case GDK_SCROLL_DOWN:
 			case GDK_SCROLL_LEFT:
-				x-=1.0;
+				x-=1.0f;
 				break;
 			default:
 				break;
